#3370 create workflow + team status directories + rename dseWorkflowStatus
