+++ wls-gui-wahllokalsystem/src/components/dse/beschlussfassung/TheBeschlussfassungStartenDialog.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script setup lang="ts">
-import type { StimmzettelerfassungTeamStatusEntry } from "@/types/dse/StimmzettelerfassungTeamStatusEntry.ts";
+import type { StimmzettelerfassungTeamStatusEntry } from "@/types/dse/stimmzettelerfassungTeamStatus/StimmzettelerfassungTeamStatusEntry.ts";
 
 import { watch } from "vue";
