--- wls-gui-wahllokalsystem/src/types/dse/StimmzettelerfassungStatus.ts
@@ -1,5 +1,0 @@
-import type { StimmzettelerfassungStatusEnum } from "@/types/dse/StimmzettelerfassungStatusEnum.ts";
-
-export interface StimmzettelerfassungStatus {
-  status: StimmzettelerfassungStatusEnum;
-}
