--- wls-gui-wahllokalsystem/src/types/dse/StimmzettelerfassungTeamStatus.ts
@@ -1,5 +1,0 @@
-import type { StimmzettelerfassungTeamStatusEnum } from "@/types/dse/StimmzettelerfassungTeamStatusEnum.ts";
-
-export interface StimmzettelerfassungTeamStatus {
-  status: StimmzettelerfassungTeamStatusEnum;
-}
