--- wls-gui-wahllokalsystem/src/types/dse/StimmzettelerfassungTeamStatusEntry.ts
@@ -1,6 +1,0 @@
-import type { StimmzettelerfassungTeamStatusEnum } from "@/types/dse/StimmzettelerfassungTeamStatusEnum.ts";
-
-export interface StimmzettelerfassungTeamStatusEntry {
-  teamID: string;
-  status: StimmzettelerfassungTeamStatusEnum;
-}
