+++ wls-gui-wahllokalsystem/src/types/dse/stimmzettelerfassungTeamStatus/StimmzettelerfassungTeamStatus.ts
@@ -1,0 +1,5 @@
+import type { StimmzettelerfassungTeamStatusEnum } from "@/types/dse/stimmzettelerfassungTeamStatus/StimmzettelerfassungTeamStatusEnum.ts";
+
+export interface StimmzettelerfassungTeamStatus {
+  status: StimmzettelerfassungTeamStatusEnum;
+}
