+++ wls-gui-wahllokalsystem/src/types/dse/stimmzettelerfassungTeamStatus/StimmzettelerfassungTeamStatusEntry.ts
@@ -1,0 +1,6 @@
+import type { StimmzettelerfassungTeamStatusEnum } from "@/types/dse/stimmzettelerfassungTeamStatus/StimmzettelerfassungTeamStatusEnum.ts";
+
+export interface StimmzettelerfassungTeamStatusEntry {
+  teamID: string;
+  status: StimmzettelerfassungTeamStatusEnum;
+}
