+++ wls-gui-wahllokalsystem/src/types/dse/stimmzettelerfassungTeamStatus/StimmzettelerfassungTeamStatusEnum.ts
