+++ wls-gui-wahllokalsystem/src/types/dse/stimmzettelerfassungWorkflowStatus/StimmzettelerfassungStatus.ts
@@ -1,0 +1,5 @@
+import type { StimmzettelerfassungStatusEnum } from "@/types/dse/stimmzettelerfassungWorkflowStatus/StimmzettelerfassungStatusEnum.ts";
+
+export interface StimmzettelerfassungStatus {
+  status: StimmzettelerfassungStatusEnum;
+}
