+++ wls-gui-wahllokalsystem/src/types/dse/stimmzettelerfassungWorkflowStatus/StimmzettelerfassungStatusEnum.ts
