+++ wls-gui-wahllokalsystem/tests/composables/navigation/navigationGuards.spec.ts
@@ -16,8 +16,8 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { useNavigationGuards } from "@/composables/navigation/navigationGuards.ts";
 import { useUserStore } from "@/stores/userStore.ts";
 import { useWorkflowStore } from "@/stores/workflowStore.ts";
-import { StimmzettelerfassungStatusEnum } from "@/types/dse/StimmzettelerfassungStatusEnum.ts";
-import { StimmzettelerfassungTeamStatusEnum } from "@/types/dse/StimmzettelerfassungTeamStatusEnum.ts";
+import { StimmzettelerfassungTeamStatusEnum } from "@/types/dse/stimmzettelerfassungTeamStatus/StimmzettelerfassungTeamStatusEnum.ts";
+import { StimmzettelerfassungStatusEnum } from "@/types/dse/stimmzettelerfassungWorkflowStatus/StimmzettelerfassungStatusEnum.ts";
 import { WahlbezirksArtEnum } from "@/types/wahlbezirksArtEnum.ts";
 
 const mockDefinitions = vi.hoisted(() => ({
@@ -25,16 +25,22 @@ const mockDefinitions = vi.hoisted(() => ({
   loadErfassungTeamStatus: vi.fn(),
 }));
 
-vi.mock("@/composables/dse/dseWorkflowStatusService.ts", () => ({
-  useDseWorkflowStatusService: () => ({
-    loadDseWorkflowStatus: mockDefinitions.loadDseWorkflowStatus,
-  }),
-}));
-vi.mock("@/composables/dse/stimmzettelerfassungTeamStatusService.ts", () => ({
-  useStimmzettelerfassungTeamStatusService: () => ({
-    loadErfassungTeamStatus: mockDefinitions.loadErfassungTeamStatus,
-  }),
-}));
+vi.mock(
+  "@/composables/dse/stimmzettelerfassungWorkflowStatus/stimmzettelerfassungStatusService.ts",
+  () => ({
+    useDseWorkflowStatusService: () => ({
+      loadDseWorkflowStatus: mockDefinitions.loadDseWorkflowStatus,
+    }),
+  })
+);
+vi.mock(
+  "@/composables/dse/stimmzettelerfassungTeamStatus/stimmzettelerfassungTeamStatusService.ts",
+  () => ({
+    useStimmzettelerfassungTeamStatusService: () => ({
+      loadErfassungTeamStatus: mockDefinitions.loadErfassungTeamStatus,
+    }),
+  })
+);
 
 const { generateRandomString } = useCommonTestDataFactory();
 const { prepareBezirkUndWahlID } = useCommonErgebnismeldungTestDataFactory();
